+++ src/components/callable-store.ts
@@ -2,7 +2,7 @@ import { Debugger } from "debug";
 import { MainFunctionMap } from "..";
 import { Transferable, TransferableValue } from "../Transferable";
 import { CallbackId, ThreadMethodKey } from "../types/general";
-import { BaseThreadMessage, CallMessage, isThreadCallbackMessage, isThreadErrorMessage, MainMessageAction } from "../types/messages";
+import { BaseThreadMessage, CallMessage, isThreadCallbackMessage, isThreadErrorMessage, isThreadFreeFunctionMessage, isThreadFunctionMessage, MainMessageAction } from "../types/messages";
 import { CountSequence, createSequence } from "../utils/sequence";
 
 export interface Callback {
@@ -13,15 +13,12 @@ export interface Callback {
 export class CallableStore {
   private callables = new Map<CallbackId, Callback>()
   private callableSequence = createSequence()
+  public mainFunctions = new Map<ThreadMethodKey, MainFunctionMap>()
+  private functionSequence = createSequence()
 
   constructor(
-    private debug: Debugger, 
-    private id: number,
-    private mainFunctions: Map<ThreadMethodKey, MainFunctionMap>,
-    private functionSequence: CountSequence
-  ) {
-
-  }
+    private debug: Debugger
+  ) {}
 
   rejectAll(withErr: Error) {
     for (const { reject } of this.callables.values()) {
@@ -67,16 +64,16 @@ export class CallableStore {
     return { msg, transferables }
   }
 
-  handleMessage(msg: BaseThreadMessage) {
+  handleMessage(msg: BaseThreadMessage, id: number): boolean {
     if (isThreadCallbackMessage(msg)) {
-      this.debug('Callable id %d resolved callback %d', this.id, msg.callbackId)
+      this.debug('Callable id %d resolved callback %d', id, msg.callbackId)
       
       const callback = this.callables.get(msg.callbackId)
       this.callables.delete(msg.callbackId)
       callback!.resolve(msg.result)
-      return
+      return true
     } else if (isThreadErrorMessage(msg)) {
-      this.debug('Callable id %d rejected callback %d', this.id, msg.callbackId)
+      this.debug('Callable id %d rejected callback %d', id, msg.callbackId)
       
       const callback = this.callables.get(msg.callbackId)
       this.callables.delete(msg.callbackId)
@@ -84,7 +81,27 @@ export class CallableStore {
       const err = new Error(msg.message)
       err.stack = msg.stack
       callback!.reject(err)
-      return
+      return true
+    } else if (isThreadFunctionMessage(msg)) {
+      this.debug('worker %d calling function %s[%d]', id, msg.key, msg.functionId)
+      
+      const fnHolder = this.mainFunctions.get(msg.key)
+      if (fnHolder) {
+        // Note: the function has to be there,
+        // if it fails, something is wrong with storing them or garbage collecting
+        const mFn = fnHolder.get(msg.functionId)!
+        mFn(...msg.args)
+      }
+      return true
+    } else if (isThreadFreeFunctionMessage(msg)) {
+      this.debug('worker %d calling free %s[%d]', id, msg.key, msg.functionId)
+      
+      const fnHolder = this.mainFunctions.get(msg.key)
+      if (fnHolder) {
+        fnHolder.delete(msg.functionId)
+      }
+      return true
     }
+    return false
   }
 }
